Refactor TKqpTasksGraph as new class (#24942)
