+++ ydb/core/kqp/executer_actor/kqp_literal_executer.cpp
@@ -76,6 +76,7 @@ public:
         : Request(std::move(request))
         , Counters(counters)
         , OwnerActor(owner)
+        , TasksGraph(Request.TxAlloc, {}, {}, Counters)
         , LiteralExecuterSpan(TWilsonKqp::LiteralExecuter, std::move(Request.TraceId), "LiteralExecuter")
         , UserRequestContext(userRequestContext)
     {
@@ -124,7 +125,7 @@ public:
 
         LOG_D("Begin literal execution, txs: " << Request.Transactions.size());
         auto& transactions = Request.Transactions;
-        FillKqpTasksGraphStages(TasksGraph, transactions);
+        TasksGraph.FillKqpTasksGraphStages(transactions);
 
         for (ui32 txIdx = 0; txIdx < transactions.size(); ++txIdx) {
             auto& tx = transactions[txIdx];
@@ -141,7 +142,7 @@ public:
             }
 
             ResponseEv->InitTxResult(tx.Body);
-            BuildKqpTaskGraphResultChannels(TasksGraph, tx.Body, txIdx);
+            TasksGraph.BuildKqpTaskGraphResultChannels(tx.Body, txIdx);
         }
 
         if (TerminateIfTimeout()) {
