+++ ydb/core/kqp/executer_actor/kqp_partition_helper.h
@@ -9,18 +9,6 @@
 
 namespace NKikimr::NKqp {
 
-struct TShardInfo {
-    struct TColumnWriteInfo {
-        ui32 MaxValueSizeBytes = 0;
-    };
-
-    TMaybe<TShardKeyRanges> KeyReadRanges;  // empty -> no reads
-    TMaybe<TShardKeyRanges> KeyWriteRanges; // empty -> no writes
-    THashMap<TString, TColumnWriteInfo> ColumnWrites;
-
-    TString ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const NScheme::TTypeRegistry& typeRegistry) const;
-};
-
 class TShardInfoWithId: public TShardInfo {
 public:
     ui64 ShardId;
@@ -36,18 +24,17 @@ struct TPhysicalShardReadSettings: public NYql::TSortingOperator<NYql::ERequestS
     NKikimr::NMiniKQL::TType* ResultType = nullptr;
 };
 
+struct TPartitionPrunerConfig {
+    TMaybe<TSerializedTableRange> BatchOperationRange;
+};
+
 class TPartitionPruner {
 public:
-    struct TConfig {
-        TMaybe<TSerializedTableRange> BatchOperationRange;
-    };
-
-public:
-    TPartitionPruner(const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv, TConfig config = {});
+    TPartitionPruner(const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv, TPartitionPrunerConfig config = {});
 
     THashMap<ui64, TShardInfo> Prune(const NKqpProto::TKqpPhyTableOperation& operation, const TStageInfo& stageInfo, bool& isFullScan);
 
-    THashMap<ui64, TShardInfo> Prune(const NKqpProto::TKqpReadRangesSource& source, const TStageInfo& stageInfo, bool& isFullScan);
+    THashMap<ui64, TShardInfo> Prune(const NKqpProto::TKqpReadRangesSource& source, const TStageInfo& stageInfo, bool& isFullScan) const;
 
     THashMap<ui64, TShardInfo> PruneEffect(const NKqpProto::TKqpPhyTableOperation& operation, const TStageInfo& stageInfo);
 
@@ -56,7 +43,7 @@ public:
 private:
     const NMiniKQL::THolderFactory* HolderFactory;
     const NMiniKQL::TTypeEnvironment* TypeEnv;
-    const TPartitionPruner::TConfig Config;
+    const TPartitionPrunerConfig Config;
 };
 
 TSerializedTableRange MakeKeyRange(const TVector<NScheme::TTypeInfo>& keyColumnTypes,
@@ -73,18 +60,18 @@ TVector<TSerializedPointOrRange> FillReadRanges(const TVector<NScheme::TTypeInfo
 
 THashMap<ui64, TShardInfo> PrunePartitions(const NKqpProto::TKqpPhyOpReadRange& readRange, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig, bool& isFullScan);
+    const TPartitionPrunerConfig& prunerConfig, bool& isFullScan);
 
 THashMap<ui64, TShardInfo> PrunePartitions(const NKqpProto::TKqpPhyOpReadRanges& readRanges, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig, bool& isFullScan);
+    const TPartitionPrunerConfig& prunerConfig, bool& isFullScan);
 
 std::pair<ui64, TShardInfo> MakeVirtualTablePartition(const NKqpProto::TKqpReadRangesSource& source, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv);
 
 THashMap<ui64, TShardInfo> PrunePartitions(const NKqpProto::TKqpReadRangesSource& source, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig, bool& isFullScan);
+    const TPartitionPrunerConfig& prunerConfig, bool& isFullScan);
 
 ui64 ExtractItemsLimit(const TStageInfo& stageInfo, const NKqpProto::TKqpPhyValue& protoItemsLimit,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv);
@@ -94,23 +81,23 @@ ui64 ExtractItemsLimit(const TStageInfo& stageInfo, const NKqpProto::TKqpPhyValu
 // partitioned and multiple ColumnShards store data from the same key range
 THashMap<ui64, TShardInfo> PrunePartitions(const NKqpProto::TKqpPhyOpReadOlapRanges& readRanges, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig, bool& isFullScan);
+    const TPartitionPrunerConfig& prunerConfig, bool& isFullScan);
 
 THashMap<ui64, TShardInfo> PrunePartitions(const NKqpProto::TKqpPhyTableOperation& operation, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig, bool& isFullScan);
+    const TPartitionPrunerConfig& prunerConfig, bool& isFullScan);
 
 THashMap<ui64, TShardInfo> PruneEffectPartitions(const NKqpProto::TKqpPhyOpUpsertRows& effect, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig);
+    const TPartitionPrunerConfig& prunerConfig);
 
 THashMap<ui64, TShardInfo> PruneEffectPartitions(const NKqpProto::TKqpPhyOpDeleteRows& effect, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig);
+    const TPartitionPrunerConfig& prunerConfig);
 
 THashMap<ui64, TShardInfo> PruneEffectPartitions(const NKqpProto::TKqpPhyTableOperation& operation, const TStageInfo& stageInfo,
     const NMiniKQL::THolderFactory& holderFactory, const NMiniKQL::TTypeEnvironment& typeEnv,
-    const TPartitionPruner::TConfig& prunerConfig);
+    const TPartitionPrunerConfig& prunerConfig);
 
 TPhysicalShardReadSettings ExtractReadSettings(
     const NKqpProto::TKqpPhyTableOperation& operation, const TStageInfo& stageInfo,
