+++ ydb/core/kqp/executer_actor/kqp_tasks_graph.cpp
@@ -1,26 +1,28 @@
 #include "kqp_tasks_graph.h"
 
+#include "kqp_partition_helper.h"
+
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/feature_flags.h>
 #include <ydb/core/base/table_index.h>
+#include <ydb/core/kqp/common/kqp_types.h>
 #include <ydb/core/kqp/common/kqp_yql.h>
+#include <ydb/core/kqp/executer_actor/kqp_executer_stats.h>
 #include <ydb/core/tx/datashard/range_ops.h>
 #include <ydb/core/tx/program/program.h>
 #include <ydb/core/tx/program/resolver.h>
 #include <ydb/core/tx/schemeshard/olap/schema/schema.h>
-
-#include <yql/essentials/core/yql_expr_optimize.h>
+#include <ydb/library/actors/core/log.h>
 #include <ydb/library/yql/dq/runtime/dq_arrow_helpers.h>
 
-#include <ydb/library/actors/core/log.h>
+#include <yql/essentials/core/yql_expr_optimize.h>
+#include <yql/essentials/providers/common/structured_token/yql_token_builder.h>
 
-namespace NKikimr {
-namespace NKqp {
+namespace NKikimr::NKqp {
 
 using namespace NYql;
 using namespace NYql::NDq;
 using namespace NYql::NNodes;
-
 
 void LogStage(const NActors::TActorContext& ctx, const TStageInfo& stageInfo) {
     LOG_DEBUG_S(ctx, NKikimrServices::KQP_EXECUTER, stageInfo.DebugString());
@@ -47,7 +49,6 @@ TTaskMeta::TReadInfo::EReadType ReadTypeFromProto(const NKqpProto::TKqpPhyOpRead
             YQL_ENSURE(false, "Invalid read type from TKqpPhyOpReadOlapRanges protobuf.");
     }
 }
-
 
 std::pair<TString, TString> SerializeKqpTasksParametersForOlap(const TStageInfo& stageInfo, const TTask& task) {
     const NKqpProto::TKqpPhyStage& stage = stageInfo.Meta.GetStage(stageInfo.Id);
@@ -90,7 +91,7 @@ std::pair<TString, TString> SerializeKqpTasksParametersForOlap(const TStageInfo&
     );
 }
 
-void FillKqpTasksGraphStages(TKqpTasksGraph& tasksGraph, const TVector<IKqpGateway::TPhysicalTxData>& txs) {
+void TKqpTasksGraph::FillKqpTasksGraphStages(const TVector<IKqpGateway::TPhysicalTxData>& txs) {
     for (size_t txIdx = 0; txIdx < txs.size(); ++txIdx) {
         auto& tx = txs[txIdx];
 
@@ -184,11 +185,10 @@ void FillKqpTasksGraphStages(TKqpTasksGraph& tasksGraph, const TVector<IKqpGatew
                 }
             }
 
-            bool stageAdded = tasksGraph.AddStageInfo(
-                TStageInfo(stageId, stage.InputsSize() + stageSourcesCount, stage.GetOutputsCount(), std::move(meta)));
+            bool stageAdded = AddStageInfo(TStageInfo(stageId, stage.InputsSize() + stageSourcesCount, stage.GetOutputsCount(), std::move(meta)));
             YQL_ENSURE(stageAdded);
 
-            auto& stageInfo = tasksGraph.GetStageInfo(stageId);
+            auto& stageInfo = GetStageInfo(stageId);
             LogStage(TlsActivationContext->AsActorContext(), stageInfo);
 
             THashSet<TTableId> tables;
@@ -228,11 +228,11 @@ void FillKqpTasksGraphStages(TKqpTasksGraph& tasksGraph, const TVector<IKqpGatew
     }
 }
 
-void BuildKqpTaskGraphResultChannels(TKqpTasksGraph& tasksGraph, const TKqpPhyTxHolder::TConstPtr& tx, ui64 txIdx) {
+void TKqpTasksGraph::BuildKqpTaskGraphResultChannels(const TKqpPhyTxHolder::TConstPtr& tx, ui64 txIdx) {
     for (ui32 i = 0; i < tx->ResultsSize(); ++i) {
         const auto& result = tx->GetResults(i);
         const auto& connection = result.GetConnection();
-        const auto& inputStageInfo = tasksGraph.GetStageInfo(TStageId(txIdx, connection.GetStageIndex()));
+        const auto& inputStageInfo = GetStageInfo(TStageId(txIdx, connection.GetStageIndex()));
         const auto& outputIdx = connection.GetOutputIndex();
 
         if (inputStageInfo.Tasks.size() < 1) {
@@ -243,14 +243,14 @@ void BuildKqpTaskGraphResultChannels(TKqpTasksGraph& tasksGraph, const TKqpPhyTx
         YQL_ENSURE(inputStageInfo.Tasks.size() == 1, "actual count: " << inputStageInfo.Tasks.size());
         auto originTaskId = inputStageInfo.Tasks[0];
 
-        auto& channel = tasksGraph.AddChannel();
+        auto& channel = AddChannel();
         channel.SrcTask = originTaskId;
         channel.SrcOutputIndex = outputIdx;
         channel.DstTask = 0;
         channel.DstInputIndex = i;
         channel.InMemory = true;
 
-        auto& originTask = tasksGraph.GetTask(originTaskId);
+        auto& originTask = GetTask(originTaskId);
 
         auto& taskOutput = originTask.Outputs[outputIdx];
         taskOutput.Type = TTaskOutputType::Map;
@@ -261,14 +261,14 @@ void BuildKqpTaskGraphResultChannels(TKqpTasksGraph& tasksGraph, const TKqpPhyTx
     }
 }
 
-void BuildTransformChannels(const TTransform& transform, const TTaskInputMeta& meta, const TString& name,
-    TKqpTasksGraph& graph, const TStageInfo& stageInfo, ui32 inputIndex,
-    const TStageInfo& inputStageInfo, ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc) {
+void TKqpTasksGraph::BuildTransformChannels(const TTransform& transform, const TTaskInputMeta& meta, const TString& name,
+    const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo, ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc)
+{
     for (ui32 taskId = 0; taskId < inputStageInfo.Tasks.size(); ++taskId) {
-        auto& originTask = graph.GetTask(inputStageInfo.Tasks[taskId]);
-        auto& targetTask = graph.GetTask(stageInfo.Tasks[taskId]);
+        auto& originTask = GetTask(inputStageInfo.Tasks[taskId]);
+        auto& targetTask = GetTask(stageInfo.Tasks[taskId]);
 
-        auto& channel = graph.AddChannel();
+        auto& channel = AddChannel();
         channel.SrcTask = originTask.Id;
         channel.SrcOutputIndex = outputIndex;
         channel.DstTask = targetTask.Id;
@@ -288,15 +288,14 @@ void BuildTransformChannels(const TTransform& transform, const TTaskInputMeta& m
     }
 }
 
-void BuildSequencerChannels(TKqpTasksGraph& graph, const TStageInfo& stageInfo, ui32 inputIndex,
-    const TStageInfo& inputStageInfo, ui32 outputIndex,
+void TKqpTasksGraph::BuildSequencerChannels(const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo, ui32 outputIndex,
     const NKqpProto::TKqpPhyCnSequencer& sequencer, bool enableSpilling, const TChannelLogFunc& logFunc)
 {
     YQL_ENSURE(stageInfo.Tasks.size() == inputStageInfo.Tasks.size());
 
-    NKikimrKqp::TKqpSequencerSettings* settings = graph.GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
+    NKikimrKqp::TKqpSequencerSettings* settings = GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
     settings->MutableTable()->CopyFrom(sequencer.GetTable());
-    settings->SetDatabase(graph.GetMeta().Database);
+    settings->SetDatabase(GetMeta().Database);
 
     const auto& tableInfo = stageInfo.Meta.TableConstInfo;
     THashSet<TString> autoIncrementColumns(sequencer.GetAutoIncrementColumns().begin(), sequencer.GetAutoIncrementColumns().end());
@@ -342,16 +341,17 @@ void BuildSequencerChannels(TKqpTasksGraph& graph, const TStageInfo& stageInfo, 
     transform.OutputType = sequencer.GetOutputType();
     TTaskInputMeta meta;
     meta.SequencerSettings = settings;
-    BuildTransformChannels(transform, meta, "Sequencer/Map", graph, stageInfo, inputIndex,
+    BuildTransformChannels(transform, meta, "Sequencer/Map", stageInfo, inputIndex,
         inputStageInfo, outputIndex, enableSpilling, logFunc);
 }
 
-void BuildChannelBetweenTasks(TKqpTasksGraph& graph, const TStageInfo& stageInfo, const TStageInfo& inputStageInfo, ui64 originTaskId,
-                              ui64 targetTaskId, ui32 inputIndex, ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc) {
-    auto& originTask = graph.GetTask(originTaskId);
-    auto& targetTask = graph.GetTask(targetTaskId);
+void TKqpTasksGraph::BuildChannelBetweenTasks(const TStageInfo& stageInfo, const TStageInfo& inputStageInfo, ui64 originTaskId,
+    ui64 targetTaskId, ui32 inputIndex, ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc)
+{
+    auto& originTask = GetTask(originTaskId);
+    auto& targetTask = GetTask(targetTaskId);
 
-    auto& channel = graph.AddChannel();
+    auto& channel = AddChannel();
     channel.SrcStageId = inputStageInfo.Id;
     channel.SrcTask = originTaskId;
     channel.SrcOutputIndex = outputIndex;
@@ -369,8 +369,9 @@ void BuildChannelBetweenTasks(TKqpTasksGraph& graph, const TStageInfo& stageInfo
     logFunc(channel.Id, originTaskId, targetTask.Id, "ParallelUnionAll/Map", !channel.InMemory);
 }
 
-void BuildParallelUnionAllChannels(TKqpTasksGraph& graph, const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo,
-                                   ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc, ui64 &nextOriginTaskId) {
+void TKqpTasksGraph::BuildParallelUnionAllChannels(const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo,
+    ui32 outputIndex, bool enableSpilling, const TChannelLogFunc& logFunc, ui64 &nextOriginTaskId)
+{
     const ui64 inputStageTasksSize = inputStageInfo.Tasks.size();
     const ui64 originStageTasksSize = stageInfo.Tasks.size();
     Y_ENSURE(originStageTasksSize);
@@ -379,18 +380,17 @@ void BuildParallelUnionAllChannels(TKqpTasksGraph& graph, const TStageInfo& stag
     for (ui64 i = 0; i < inputStageTasksSize; ++i) {
         const auto originTaskId = inputStageInfo.Tasks[i];
         const auto targetTaskId = stageInfo.Tasks[nextOriginTaskId];
-        BuildChannelBetweenTasks(graph, stageInfo, inputStageInfo, originTaskId, targetTaskId, inputIndex, outputIndex, enableSpilling, logFunc);
+        BuildChannelBetweenTasks(stageInfo, inputStageInfo, originTaskId, targetTaskId, inputIndex, outputIndex, enableSpilling, logFunc);
         nextOriginTaskId = (nextOriginTaskId + 1) % originStageTasksSize;
     }
 }
 
-void BuildStreamLookupChannels(TKqpTasksGraph& graph, const TStageInfo& stageInfo, ui32 inputIndex,
-    const TStageInfo& inputStageInfo, ui32 outputIndex,
+void TKqpTasksGraph::BuildStreamLookupChannels(const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo, ui32 outputIndex,
     const NKqpProto::TKqpPhyCnStreamLookup& streamLookup, bool enableSpilling, const TChannelLogFunc& logFunc)
 {
     YQL_ENSURE(stageInfo.Tasks.size() == inputStageInfo.Tasks.size());
 
-    NKikimrKqp::TKqpStreamLookupSettings* settings = graph.GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
+    NKikimrKqp::TKqpStreamLookupSettings* settings = GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
 
     settings->MutableTable()->CopyFrom(streamLookup.GetTable());
 
@@ -435,7 +435,7 @@ void BuildStreamLookupChannels(TKqpTasksGraph& graph, const TStageInfo& stageInf
     settings->SetAllowNullKeysPrefixSize(streamLookup.GetAllowNullKeysPrefixSize());
 
     if (streamLookup.GetIsTableImmutable()
-        && graph.GetMeta().RequestIsolationLevel == NKikimrKqp::EIsolationLevel::ISOLATION_LEVEL_READ_STALE)
+        && GetMeta().RequestIsolationLevel == NKikimrKqp::EIsolationLevel::ISOLATION_LEVEL_READ_STALE)
     {
         settings->SetAllowUseFollowers(true);
         settings->SetIsTableImmutable(true);
@@ -447,17 +447,16 @@ void BuildStreamLookupChannels(TKqpTasksGraph& graph, const TStageInfo& stageInf
     streamLookupTransform.OutputType = streamLookup.GetResultType();
     TTaskInputMeta meta;
     meta.StreamLookupSettings = settings;
-    BuildTransformChannels(streamLookupTransform, meta, "StreamLookup/Map", graph, stageInfo, inputIndex,
+    BuildTransformChannels(streamLookupTransform, meta, "StreamLookup/Map", stageInfo, inputIndex,
         inputStageInfo, outputIndex, enableSpilling, logFunc);
 }
 
-void BuildVectorResolveChannels(TKqpTasksGraph& graph, const TStageInfo& stageInfo, ui32 inputIndex,
-    const TStageInfo& inputStageInfo, ui32 outputIndex,
+void TKqpTasksGraph::BuildVectorResolveChannels(const TStageInfo& stageInfo, ui32 inputIndex, const TStageInfo& inputStageInfo, ui32 outputIndex,
     const NKqpProto::TKqpPhyCnVectorResolve& vectorResolve, bool enableSpilling, const TChannelLogFunc& logFunc)
 {
     YQL_ENSURE(stageInfo.Tasks.size() == inputStageInfo.Tasks.size());
 
-    NKikimrTxDataShard::TKqpVectorResolveSettings* settings = graph.GetMeta().Allocate<NKikimrTxDataShard::TKqpVectorResolveSettings>();
+    auto* settings = GetMeta().Allocate<NKikimrTxDataShard::TKqpVectorResolveSettings>();
 
     *settings->MutableIndexSettings() = vectorResolve.GetIndexSettings();
 
@@ -509,20 +508,18 @@ void BuildVectorResolveChannels(TKqpTasksGraph& graph, const TStageInfo& stageIn
     vectorResolveTransform.OutputType = vectorResolve.GetOutputType();
     TTaskInputMeta meta;
     meta.VectorResolveSettings = settings;
-    BuildTransformChannels(vectorResolveTransform, meta, "VectorResolve/Map", graph, stageInfo, inputIndex,
+    BuildTransformChannels(vectorResolveTransform, meta, "VectorResolve/Map", stageInfo, inputIndex,
         inputStageInfo, outputIndex, enableSpilling, logFunc);
 }
 
-void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
-    ui64 txId, bool enableSpilling, bool enableShuffleElimination)
-{
+void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, bool enableSpilling, bool enableShuffleElimination) {
     auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
 
     if (stage.GetIsEffectsStage() && stage.GetSinks().empty()) {
         YQL_ENSURE(stageInfo.OutputsCount == 1);
 
         for (auto& taskId : stageInfo.Tasks) {
-            auto& task = tasksGraph.GetTask(taskId);
+            auto& task = GetTask(taskId);
             auto& taskOutput = task.Outputs[0];
             taskOutput.Type = TTaskOutputType::Effects;
         }
@@ -541,7 +538,7 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
     if (enableShuffleElimination && !isFusedWithScanStage) { // taskIdHash can be already set if it is a fused stage, so hashpartition will derive columnv1 parameters from there
         for (ui32 inputIndex = 0; inputIndex < stage.InputsSize(); ++inputIndex) {
             const auto& input = stage.GetInputs(inputIndex);
-            auto& originStageInfo = tasksGraph.GetStageInfo(NYql::NDq::TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
+            auto& originStageInfo = GetStageInfo(NYql::NDq::TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
             ui32 outputIdx = input.GetOutputIndex();
             columnShardHashV1Params = originStageInfo.Meta.GetColumnShardHashV1Params(outputIdx);
             if (input.GetTypeCase() == NKqpProto::TKqpPhyConnection::kMap || inputIndex == stage.InputsSize() - 1) { // this branch is only for logging purposes
@@ -602,12 +599,12 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
     ui64 nextOriginTaskId = 0;
     for (auto& input : stage.GetInputs()) {
         ui32 inputIdx = input.GetInputIndex();
-        auto& inputStageInfo = tasksGraph.GetStageInfo(TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
+        auto& inputStageInfo = GetStageInfo(TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
         const auto& outputIdx = input.GetOutputIndex();
 
         switch (input.GetTypeCase()) {
             case NKqpProto::TKqpPhyConnection::kUnionAll:
-                BuildUnionAllChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
+                BuildUnionAllChannels(*this, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
                 break;
             case NKqpProto::TKqpPhyConnection::kHashShuffle: {
                 std::optional<EHashShuffleFuncType> hashKind;
@@ -653,7 +650,7 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
 
                 Y_ENSURE(hashKind.has_value(), "HashKind wasn't set!");
                 BuildHashShuffleChannels(
-                    tasksGraph,
+                    *this,
                     stageInfo,
                     inputIdx,
                     inputStageInfo,
@@ -667,10 +664,10 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
                 break;
             }
             case NKqpProto::TKqpPhyConnection::kBroadcast:
-                BuildBroadcastChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
+                BuildBroadcastChannels(*this, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
                 break;
             case NKqpProto::TKqpPhyConnection::kMap:
-                BuildMapChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
+                BuildMapChannels(*this, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log);
                 break;
             case NKqpProto::TKqpPhyConnection::kMerge: {
                 TVector<TSortColumn> sortColumns;
@@ -682,29 +679,28 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
                     );
                 }
                 // TODO: spilling?
-                BuildMergeChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx, sortColumns, log);
+                BuildMergeChannels(*this, stageInfo, inputIdx, inputStageInfo, outputIdx, sortColumns, log);
                 break;
             }
             case NKqpProto::TKqpPhyConnection::kSequencer: {
-                BuildSequencerChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx,
+                BuildSequencerChannels(stageInfo, inputIdx, inputStageInfo, outputIdx,
                     input.GetSequencer(), enableSpilling, log);
                 break;
             }
 
             case NKqpProto::TKqpPhyConnection::kStreamLookup: {
-                BuildStreamLookupChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx,
+                BuildStreamLookupChannels(stageInfo, inputIdx, inputStageInfo, outputIdx,
                     input.GetStreamLookup(), enableSpilling, log);
                 break;
             }
 
             case NKqpProto::TKqpPhyConnection::kParallelUnionAll: {
-                BuildParallelUnionAllChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log, nextOriginTaskId);
+                BuildParallelUnionAllChannels(stageInfo, inputIdx, inputStageInfo, outputIdx, enableSpilling, log, nextOriginTaskId);
                 break;
             }
 
             case NKqpProto::TKqpPhyConnection::kVectorResolve: {
-                BuildVectorResolveChannels(tasksGraph, stageInfo, inputIdx, inputStageInfo, outputIdx,
-                    input.GetVectorResolve(), enableSpilling, log);
+                BuildVectorResolveChannels(stageInfo, inputIdx, inputStageInfo, outputIdx, input.GetVectorResolve(), enableSpilling, log);
                 break;
             }
 
@@ -714,19 +710,19 @@ void BuildKqpStageChannels(TKqpTasksGraph& tasksGraph, TStageInfo& stageInfo,
     }
 }
 
-bool IsCrossShardChannel(const TKqpTasksGraph& tasksGraph, const TChannel& channel) {
+bool TKqpTasksGraph::IsCrossShardChannel(const TChannel& channel) const {
     YQL_ENSURE(channel.SrcTask);
 
     if (!channel.DstTask) {
         return false;
     }
 
-    ui64 targetShard = tasksGraph.GetTask(channel.DstTask).Meta.ShardId;
+    ui64 targetShard = GetTask(channel.DstTask).Meta.ShardId;
     if (!targetShard) {
         return false;
     }
 
-    ui64 srcShard = tasksGraph.GetTask(channel.SrcTask).Meta.ShardId;
+    ui64 srcShard = GetTask(channel.SrcTask).Meta.ShardId;
     return srcShard && targetShard != srcShard;
 }
 
@@ -782,9 +778,7 @@ void TShardKeyRanges::MakeFull(TSerializedPointOrRange&& pointOrRange) {
     }
 }
 
-
 void TShardKeyRanges::MergeWritePoints(TShardKeyRanges&& other, const TVector<NScheme::TTypeInfo>& keyTypes) {
-
     if (IsFullRange()) {
         return;
     }
@@ -844,8 +838,7 @@ void TShardKeyRanges::MergeWritePoints(TShardKeyRanges&& other, const TVector<NS
     Ranges = std::move(result);
 }
 
-TString TShardKeyRanges::ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const NScheme::TTypeRegistry& typeRegistry) const
-{
+TString TShardKeyRanges::ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const NScheme::TTypeRegistry& typeRegistry) const {
     TStringBuilder sb;
     sb << "TShardKeyRanges{ ";
     if (IsFullRange()) {
@@ -978,8 +971,9 @@ void FillEndpointDesc(NDqProto::TEndpoint& endpoint, const TTask& task) {
     }
 }
 
-void FillChannelDesc(const TKqpTasksGraph& tasksGraph, NDqProto::TChannel& channelDesc, const TChannel& channel,
-    const NKikimrConfig::TTableServiceConfig::EChannelTransportVersion chanTransportVersion, bool enableSpilling) {
+void TKqpTasksGraph::FillChannelDesc(NDqProto::TChannel& channelDesc, const TChannel& channel,
+    const NKikimrConfig::TTableServiceConfig::EChannelTransportVersion chanTransportVersion, bool enableSpilling) const
+{
     channelDesc.SetId(channel.Id);
     channelDesc.SetSrcStageId(channel.SrcStageId.StageId);
     channelDesc.SetDstStageId(channel.DstStageId.StageId);
@@ -987,14 +981,14 @@ void FillChannelDesc(const TKqpTasksGraph& tasksGraph, NDqProto::TChannel& chann
     channelDesc.SetDstTaskId(channel.DstTask);
     channelDesc.SetEnableSpilling(enableSpilling);
 
-    const auto& resultChannelProxies = tasksGraph.GetMeta().ResultChannelProxies;
+    const auto& resultChannelProxies = GetMeta().ResultChannelProxies;
 
     YQL_ENSURE(channel.SrcTask);
-    const auto& srcTask = tasksGraph.GetTask(channel.SrcTask);
+    const auto& srcTask = GetTask(channel.SrcTask);
     FillEndpointDesc(*channelDesc.MutableSrcEndpoint(), srcTask);
 
     if (channel.DstTask) {
-        FillEndpointDesc(*channelDesc.MutableDstEndpoint(), tasksGraph.GetTask(channel.DstTask));
+        FillEndpointDesc(*channelDesc.MutableDstEndpoint(), GetTask(channel.DstTask));
     } else if (!resultChannelProxies.empty()) {
         auto it = resultChannelProxies.find(channel.Id);
         YQL_ENSURE(it != resultChannelProxies.end());
@@ -1004,7 +998,7 @@ void FillChannelDesc(const TKqpTasksGraph& tasksGraph, NDqProto::TChannel& chann
         ActorIdToProto(srcTask.Meta.ExecuterId, channelDesc.MutableDstEndpoint()->MutableActorId());
     }
 
-    channelDesc.SetIsPersistent(IsCrossShardChannel(tasksGraph, channel));
+    channelDesc.SetIsPersistent(IsCrossShardChannel(channel));
     channelDesc.SetInMemory(channel.InMemory);
     if (chanTransportVersion == NKikimrConfig::TTableServiceConfig::CTV_OOB_PICKLE_1_0) {
         channelDesc.SetTransportVersion(NDqProto::EDataTransportVersion::DATA_TRANSPORT_OOB_PICKLE_1_0);
@@ -1180,14 +1174,9 @@ void FillTaskMeta(const TStageInfo& stageInfo, const TTask& task, NYql::NDqProto
     }
 }
 
-void FillOutputDesc(
-    const TKqpTasksGraph& tasksGraph,
-    NYql::NDqProto::TTaskOutput& outputDesc,
-    const TTaskOutput& output,
-    ui32 outputIdx,
-    bool enableSpilling,
-    const TStageInfo& stageInfo
-) {
+void TKqpTasksGraph::FillOutputDesc(NYql::NDqProto::TTaskOutput& outputDesc, const TTaskOutput& output, ui32 outputIdx,
+    bool enableSpilling, const TStageInfo& stageInfo) const
+{
     switch (output.Type) {
         case TTaskOutputType::Map:
             YQL_ENSURE(output.Channels.size() == 1);
@@ -1297,13 +1286,13 @@ void FillOutputDesc(
 
     for (auto& channel : output.Channels) {
         auto& channelDesc = *outputDesc.AddChannels();
-        FillChannelDesc(tasksGraph, channelDesc, tasksGraph.GetChannel(channel), tasksGraph.GetMeta().ChannelTransportVersion, enableSpilling);
+        FillChannelDesc(channelDesc, GetChannel(channel), GetMeta().ChannelTransportVersion, enableSpilling);
     }
 }
 
-void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput& inputDesc, const TTaskInput& input, bool serializeAsyncIoSettings, bool& enableMetering) {
-    const auto& snapshot = tasksGraph.GetMeta().Snapshot;
-    const auto& lockTxId = tasksGraph.GetMeta().LockTxId;
+void TKqpTasksGraph::FillInputDesc(NYql::NDqProto::TTaskInput& inputDesc, const TTaskInput& input, bool serializeAsyncIoSettings, bool& enableMetering) const {
+    const auto& snapshot = GetMeta().Snapshot;
+    const auto& lockTxId = GetMeta().LockTxId;
 
     switch (input.Type()) {
         case NYql::NDq::TTaskInputType::Source:
@@ -1319,8 +1308,8 @@ void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput&
                     input.Meta.SourceSettings->MutableSnapshot()->SetTxId(snapshot.TxId);
                 }
 
-                if (tasksGraph.GetMeta().UseFollowers || isTableImmutable) {
-                    input.Meta.SourceSettings->SetUseFollowers(tasksGraph.GetMeta().UseFollowers || isTableImmutable);
+                if (GetMeta().UseFollowers || isTableImmutable) {
+                    input.Meta.SourceSettings->SetUseFollowers(GetMeta().UseFollowers || isTableImmutable);
                 }
 
                 if (serializeAsyncIoSettings) {
@@ -1358,7 +1347,7 @@ void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput&
 
     for (ui64 channel : input.Channels) {
         auto& channelDesc = *inputDesc.AddChannels();
-        FillChannelDesc(tasksGraph, channelDesc, tasksGraph.GetChannel(channel), tasksGraph.GetMeta().ChannelTransportVersion, false);
+        FillChannelDesc(channelDesc, GetChannel(channel), GetMeta().ChannelTransportVersion, false);
     }
 
     if (input.Transform) {
@@ -1370,23 +1359,23 @@ void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput&
             enableMetering = true;
             YQL_ENSURE(input.Meta.StreamLookupSettings);
             bool isTableImmutable = input.Meta.StreamLookupSettings->GetIsTableImmutable() &&
-                tasksGraph.GetMeta().RequestIsolationLevel == NKikimrKqp::EIsolationLevel::ISOLATION_LEVEL_READ_STALE;
+                GetMeta().RequestIsolationLevel == NKikimrKqp::EIsolationLevel::ISOLATION_LEVEL_READ_STALE;
 
             if (snapshot.IsValid() && !isTableImmutable) {
                 input.Meta.StreamLookupSettings->MutableSnapshot()->SetStep(snapshot.Step);
                 input.Meta.StreamLookupSettings->MutableSnapshot()->SetTxId(snapshot.TxId);
             } else {
-                YQL_ENSURE(tasksGraph.GetMeta().AllowInconsistentReads || isTableImmutable, "Expected valid snapshot or enabled inconsistent read mode");
+                YQL_ENSURE(GetMeta().AllowInconsistentReads || isTableImmutable, "Expected valid snapshot or enabled inconsistent read mode");
                 input.Meta.StreamLookupSettings->SetAllowInconsistentReads(true);
             }
 
             if (lockTxId && !isTableImmutable) {
                 input.Meta.StreamLookupSettings->SetLockTxId(*lockTxId);
-                input.Meta.StreamLookupSettings->SetLockNodeId(tasksGraph.GetMeta().LockNodeId);
+                input.Meta.StreamLookupSettings->SetLockNodeId(GetMeta().LockNodeId);
             }
 
-            if (tasksGraph.GetMeta().LockMode && !isTableImmutable) {
-                input.Meta.StreamLookupSettings->SetLockMode(*tasksGraph.GetMeta().LockMode);
+            if (GetMeta().LockMode && !isTableImmutable) {
+                input.Meta.StreamLookupSettings->SetLockMode(*GetMeta().LockMode);
             }
 
             transformProto->MutableSettings()->PackFrom(*input.Meta.StreamLookupSettings);
@@ -1402,11 +1391,11 @@ void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput&
 
             if (lockTxId) {
                 input.Meta.VectorResolveSettings->SetLockTxId(*lockTxId);
-                input.Meta.VectorResolveSettings->SetLockNodeId(tasksGraph.GetMeta().LockNodeId);
+                input.Meta.VectorResolveSettings->SetLockNodeId(GetMeta().LockNodeId);
             }
 
-            if (tasksGraph.GetMeta().LockMode) {
-                input.Meta.VectorResolveSettings->SetLockMode(*tasksGraph.GetMeta().LockMode);
+            if (GetMeta().LockMode) {
+                input.Meta.VectorResolveSettings->SetLockMode(*GetMeta().LockMode);
             }
 
             transformProto->MutableSettings()->PackFrom(*input.Meta.VectorResolveSettings);
@@ -1414,12 +1403,8 @@ void FillInputDesc(const TKqpTasksGraph& tasksGraph, NYql::NDqProto::TTaskInput&
     }
 }
 
-void SerializeTaskToProto(
-        const TKqpTasksGraph& tasksGraph,
-        const TTask& task,
-        NYql::NDqProto::TDqTask* result,
-        bool serializeAsyncIoSettings) {
-    auto& stageInfo = tasksGraph.GetStageInfo(task.StageId);
+void TKqpTasksGraph::SerializeTaskToProto(const TTask& task, NYql::NDqProto::TDqTask* result, bool serializeAsyncIoSettings) const {
+    auto& stageInfo = GetStageInfo(task.StageId);
     ActorIdToProto(task.Meta.ExecuterId, result->MutableExecuter()->MutableActorId());
     result->SetId(task.Id);
     result->SetStageId(stageInfo.Id.StageId);
@@ -1443,16 +1428,16 @@ void SerializeTaskToProto(
     }
 
     for (const auto& input : task.Inputs) {
-        FillInputDesc(tasksGraph, *result->AddInputs(), input, serializeAsyncIoSettings, enableMetering);
+        FillInputDesc(*result->AddInputs(), input, serializeAsyncIoSettings, enableMetering);
     }
 
     bool enableSpilling = false;
     if (task.Outputs.size() > 1) {
-        enableSpilling = tasksGraph.GetMeta().AllowWithSpilling;
+        enableSpilling = GetMeta().AllowWithSpilling;
     }
     for (ui32 outputIdx = 0; outputIdx < task.Outputs.size(); ++outputIdx) {
         const auto& output = task.Outputs[outputIdx];
-        FillOutputDesc(tasksGraph, *result->AddOutputs(), output, outputIdx, enableSpilling, stageInfo);
+        FillOutputDesc(*result->AddOutputs(), output, outputIdx, enableSpilling, stageInfo);
     }
 
     const NKqpProto::TKqpPhyStage& stage = stageInfo.Meta.GetStage(stageInfo.Id);
@@ -1467,30 +1452,30 @@ void SerializeTaskToProto(
         }
     }
 
-    SerializeCtxToMap(*tasksGraph.GetMeta().UserRequestContext, *result->MutableRequestContext());
+    SerializeCtxToMap(*GetMeta().UserRequestContext, *result->MutableRequestContext());
 
     result->SetDisableMetering(!enableMetering);
-    result->SetCreateSuspended(tasksGraph.GetMeta().CreateSuspended);
+    result->SetCreateSuspended(GetMeta().CreateSuspended);
     FillTaskMeta(stageInfo, task, *result);
 }
 
-NYql::NDqProto::TDqTask* ArenaSerializeTaskToProto(TKqpTasksGraph& tasksGraph, const TTask& task, bool serializeAsyncIoSettings) {
-    NYql::NDqProto::TDqTask* result = tasksGraph.GetMeta().Allocate<NYql::NDqProto::TDqTask>();
-    SerializeTaskToProto(tasksGraph, task, result, serializeAsyncIoSettings);
+NYql::NDqProto::TDqTask* TKqpTasksGraph::ArenaSerializeTaskToProto(const TTask& task, bool serializeAsyncIoSettings) {
+    NYql::NDqProto::TDqTask* result = GetMeta().Allocate<NYql::NDqProto::TDqTask>();
+    SerializeTaskToProto(task, result, serializeAsyncIoSettings);
     return result;
 }
 
-void PersistTasksGraphInfo(const TKqpTasksGraph& tasksGraph, NKikimrKqp::TQueryPhysicalGraph& result) {
+void TKqpTasksGraph::PersistTasksGraphInfo(NKikimrKqp::TQueryPhysicalGraph& result) const {
     auto& resultTasks = *result.MutableTasks();
 
-    const auto& tasks = tasksGraph.GetTasks();
+    const auto& tasks = GetTasks();
     resultTasks.Reserve(tasks.size());
     for (const auto& task : tasks) {
         auto& resultTask = *resultTasks.Add();
         resultTask.SetTxId(task.StageId.TxId);
 
         auto* taskInfo = resultTask.MutableDqTask();
-        SerializeTaskToProto(tasksGraph, task, taskInfo, false);
+        SerializeTaskToProto(task, taskInfo, false);
 
         taskInfo->ClearProgram();
         taskInfo->ClearSecureParams();
@@ -1499,7 +1484,7 @@ void PersistTasksGraphInfo(const TKqpTasksGraph& tasksGraph, NKikimrKqp::TQueryP
 
 // Restored graph only requires to update authentication secrets
 // and to reassign existing tasks between actual nodes.
-void RestoreTasksGraphInfo(TKqpTasksGraph& tasksGraph, const NKikimrKqp::TQueryPhysicalGraph& graphInfo) {
+void TKqpTasksGraph::RestoreTasksGraphInfo(const NKikimrKqp::TQueryPhysicalGraph& graphInfo) {
     const auto restoreDqTransform = [](const auto& protoInfo) -> TMaybe<TTransform> {
         if (!protoInfo.HasTransform()) {
             return Nothing();
@@ -1536,8 +1521,8 @@ void RestoreTasksGraphInfo(TKqpTasksGraph& tasksGraph, const NKikimrKqp::TQueryP
         const auto txId = task.GetTxId();
         const auto& taskInfo = task.GetDqTask();
 
-        auto& stageInfo = tasksGraph.GetStageInfo({txId, taskInfo.GetStageId()});
-        auto& newTask = tasksGraph.AddTask(stageInfo);
+        auto& stageInfo = GetStageInfo({txId, taskInfo.GetStageId()});
+        auto& newTask = AddTask(stageInfo);
         YQL_ENSURE(taskInfo.GetId() == newTask.Id);
         newTask.Meta.TaskParams.insert(taskInfo.GetTaskParams().begin(), taskInfo.GetTaskParams().end());
         newTask.Meta.ReadRanges.assign(taskInfo.GetReadRanges().begin(), taskInfo.GetReadRanges().end());
@@ -1658,13 +1643,1069 @@ void RestoreTasksGraphInfo(TKqpTasksGraph& tasksGraph, const NKikimrKqp::TQueryP
     }
 
     for (const auto& [id, channel] : channels) {
-        auto& newChannel = tasksGraph.AddChannel();
+        auto& newChannel = AddChannel();
         newChannel = channel;
         YQL_ENSURE(id == newChannel.Id);
     }
 
-    tasksGraph.GetMeta().IsRestored = true;
+    GetMeta().IsRestored = true;
 }
+
+void TKqpTasksGraph::BuildSysViewScanTasks(TStageInfo& stageInfo) {
+    Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.IsSysView());
+
+    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    const auto& holderFactory = TxAlloc->HolderFactory;
+    const auto& typeEnv = TxAlloc->TypeEnv;
+    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
+    const auto& keyTypes = tableInfo->KeyColumnTypes;
+
+    for (auto& op : stage.GetTableOps()) {
+        Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.TablePath == op.GetTable().GetPath());
+
+        auto& task = AddTask(stageInfo);
+        TShardKeyRanges keyRanges;
+
+        switch (op.GetTypeCase()) {
+            case NKqpProto::TKqpPhyTableOperation::kReadRange:
+                stageInfo.Meta.SkipNullKeys.assign(
+                    op.GetReadRange().GetSkipNullKeys().begin(),
+                    op.GetReadRange().GetSkipNullKeys().end()
+                );
+                keyRanges.Add(MakeKeyRange(
+                    keyTypes, op.GetReadRange().GetKeyRange(),
+                    stageInfo, holderFactory, typeEnv)
+                );
+                break;
+            case NKqpProto::TKqpPhyTableOperation::kReadRanges:
+                keyRanges.CopyFrom(FillReadRanges(keyTypes, op.GetReadRanges(), stageInfo, typeEnv));
+                break;
+            default:
+                YQL_ENSURE(false, "Unexpected table scan operation: " << (ui32) op.GetTypeCase());
+        }
+
+        TTaskMeta::TShardReadInfo readInfo = {
+            .Ranges = std::move(keyRanges),
+            .Columns = BuildKqpColumns(op, tableInfo),
+        };
+
+        auto readSettings = ExtractReadSettings(op, stageInfo, holderFactory, typeEnv);
+        task.Meta.Reads.ConstructInPlace();
+        task.Meta.Reads->emplace_back(std::move(readInfo));
+        task.Meta.ReadInfo.SetSorting(readSettings.GetSorting());
+        task.Meta.Type = TTaskMeta::TTaskType::Compute;
+
+        // TODO: LOG_D("Stage " << stageInfo.Id << " create sysview scan task: " << task.Id);
+    }
+}
+
+ui32 TKqpTasksGraph::GetMaxTasksAggregation(TStageInfo& stageInfo, const ui32 previousTasksCount, const ui32 nodesCount) {
+    auto& intros = stageInfo.Introspections;
+    if (AggregationSettings.HasAggregationComputeThreads()) {
+        auto threads = AggregationSettings.GetAggregationComputeThreads();
+        intros.push_back("Considering AggregationComputeThreads value - " + ToString(threads));
+        return std::max<ui32>(1, threads);
+    } else if (nodesCount) {
+        const TStagePredictor& predictor = stageInfo.Meta.Tx.Body->GetCalculationPredictor(stageInfo.Id.StageId);
+        auto result = predictor.CalcTasksOptimalCount(TStagePredictor::GetUsableThreads(), previousTasksCount / nodesCount, intros) * nodesCount;
+        intros.push_back("Predicted value for aggregation - " + ToString(result));
+        return result;
+    } else {
+        intros.push_back("Unknown nodes count for aggregation - using value 1");
+        return 1;
+    }
+}
+
+bool TKqpTasksGraph::BuildComputeTasks(TStageInfo& stageInfo, const ui32 nodesCount) {
+    auto& intros = stageInfo.Introspections;
+    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    // TODO: move outside
+    if (GetMeta().IsRestored) {
+        for (const auto taskId : stageInfo.Tasks) {
+            auto& task = GetTask(taskId);
+            task.Meta.Type = TTaskMeta::TTaskType::Compute;
+        }
+        return false;
+    }
+
+    bool unknownAffectedShardCount = false;
+    ui32 partitionsCount = 1;
+    ui32 inputTasks = 0;
+    bool isShuffle = false;
+    bool forceMapTasks = false;
+    bool isParallelUnionAll = false;
+    ui32 mapConnectionCount = 0;
+
+    for (ui32 inputIndex = 0; inputIndex < stage.InputsSize(); ++inputIndex) {
+        const auto& input = stage.GetInputs(inputIndex);
+
+        // Current assumptions:
+        // 1. All stage's inputs, except 1st one, must be a `Broadcast` or `UnionAll`
+        // 2. Stages where 1st input is `Broadcast` are not partitioned.
+        if (inputIndex > 0) {
+            switch (input.GetTypeCase()) {
+                case NKqpProto::TKqpPhyConnection::kBroadcast:
+                case NKqpProto::TKqpPhyConnection::kHashShuffle:
+                case NKqpProto::TKqpPhyConnection::kUnionAll:
+                case NKqpProto::TKqpPhyConnection::kMerge:
+                case NKqpProto::TKqpPhyConnection::kStreamLookup:
+                case NKqpProto::TKqpPhyConnection::kMap:
+                case NKqpProto::TKqpPhyConnection::kParallelUnionAll:
+                case NKqpProto::TKqpPhyConnection::kVectorResolve:
+                    break;
+                default:
+                    YQL_ENSURE(false, "Unexpected connection type: " << (ui32)input.GetTypeCase() << Endl);
+                    // TODO: << this->DebugString());
+            }
+        }
+
+        auto& originStageInfo = GetStageInfo(NYql::NDq::TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
+
+        switch (input.GetTypeCase()) {
+            case NKqpProto::TKqpPhyConnection::kHashShuffle: {
+                inputTasks += originStageInfo.Tasks.size();
+                isShuffle = true;
+                break;
+            }
+            case NKqpProto::TKqpPhyConnection::kStreamLookup: {
+                partitionsCount = originStageInfo.Tasks.size();
+                unknownAffectedShardCount = true;
+                intros.push_back("Resetting compute tasks count because input " + ToString(inputIndex) + " is StreamLookup - " + ToString(partitionsCount));
+                break;
+            }
+            case NKqpProto::TKqpPhyConnection::kMap: {
+                partitionsCount = originStageInfo.Tasks.size();
+                forceMapTasks = true;
+                ++mapConnectionCount;
+                intros.push_back("Resetting compute tasks count because input " + ToString(inputIndex) + " is Map - " + ToString(partitionsCount));
+                break;
+            }
+            case NKqpProto::TKqpPhyConnection::kParallelUnionAll: {
+                inputTasks += originStageInfo.Tasks.size();
+                isParallelUnionAll = true;
+                break;
+            }
+            case NKqpProto::TKqpPhyConnection::kVectorResolve: {
+                partitionsCount = originStageInfo.Tasks.size();
+                unknownAffectedShardCount = true;
+                intros.push_back("Resetting compute tasks count because input " + ToString(inputIndex) + " is VectorResolve - " + ToString(partitionsCount));
+                break;
+            }
+            default:
+                break;
+        }
+
+    }
+
+    Y_ENSURE(mapConnectionCount <= 1, "Only a single map connection is allowed");
+
+    if ((isShuffle || isParallelUnionAll) && !forceMapTasks) {
+        if (stage.GetTaskCount()) {
+            partitionsCount = stage.GetTaskCount();
+            intros.push_back("Manually overridden - " + ToString(partitionsCount));
+        } else {
+            partitionsCount = std::max(partitionsCount, GetMaxTasksAggregation(stageInfo, inputTasks, nodesCount));
+        }
+    }
+
+    intros.push_back("Actual number of compute tasks - " + ToString(partitionsCount));
+
+    for (ui32 i = 0; i < partitionsCount; ++i) {
+        auto& task = AddTask(stageInfo);
+        task.Meta.Type = TTaskMeta::TTaskType::Compute;
+        // TODO: LOG_D("Stage " << stageInfo.Id << " create compute task: " << task.Id);
+    }
+
+    return unknownAffectedShardCount;
+}
+
+void FillReadInfo(TTaskMeta& taskMeta, ui64 itemsLimit, const NYql::ERequestSorting sorting) {
+    if (taskMeta.Reads && !taskMeta.Reads.GetRef().empty()) {
+        // Validate parameters
+        YQL_ENSURE(taskMeta.ReadInfo.ItemsLimit == itemsLimit);
+        YQL_ENSURE(taskMeta.ReadInfo.GetSorting() == sorting);
+        return;
+    }
+
+    taskMeta.ReadInfo.ItemsLimit = itemsLimit;
+    taskMeta.ReadInfo.SetSorting(sorting);
+    taskMeta.ReadInfo.ReadType = TTaskMeta::TReadInfo::EReadType::Rows;
+}
+
+TTaskMeta::TReadInfo::EReadType OlapReadTypeFromProto(const NKqpProto::TKqpPhyOpReadOlapRanges::EReadType& type) {
+    switch (type) {
+        case NKqpProto::TKqpPhyOpReadOlapRanges::ROWS:
+            return TTaskMeta::TReadInfo::EReadType::Rows;
+        case NKqpProto::TKqpPhyOpReadOlapRanges::BLOCKS:
+            return TTaskMeta::TReadInfo::EReadType::Blocks;
+        default:
+            YQL_ENSURE(false, "Invalid read type from TKqpPhyOpReadOlapRanges protobuf.");
+    }
+}
+
+void FillOlapReadInfo(TTaskMeta& taskMeta, NKikimr::NMiniKQL::TType* resultType, const TMaybe<::NKqpProto::TKqpPhyOpReadOlapRanges>& readOlapRange) {
+    if (taskMeta.Reads && !taskMeta.Reads.GetRef().empty()) {
+        // Validate parameters
+        if (!readOlapRange || readOlapRange->GetOlapProgram().empty()) {
+            YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program.empty());
+            return;
+        }
+
+        YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program == readOlapRange->GetOlapProgram());
+        return;
+    }
+
+    if (resultType) {
+        YQL_ENSURE(resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Struct
+            || resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Tuple);
+
+        auto* resultStructType = static_cast<NKikimr::NMiniKQL::TStructType*>(resultType);
+        ui32 resultColsCount = resultStructType->GetMembersCount();
+
+        taskMeta.ReadInfo.ResultColumnsTypes.reserve(resultColsCount);
+        for (ui32 i = 0; i < resultColsCount; ++i) {
+            taskMeta.ReadInfo.ResultColumnsTypes.emplace_back();
+            auto memberType = resultStructType->GetMemberType(i);
+            NScheme::TTypeInfo typeInfo = NScheme::TypeInfoFromMiniKQLType(memberType);
+            taskMeta.ReadInfo.ResultColumnsTypes.back() = typeInfo;
+        }
+    }
+    if (!readOlapRange || readOlapRange->GetOlapProgram().empty()) {
+        return;
+    }
+    {
+        Y_ABORT_UNLESS(taskMeta.ReadInfo.GroupByColumnNames.empty());
+        std::vector<std::string> groupByColumns;
+        for (auto&& i : readOlapRange->GetGroupByColumnNames()) {
+            groupByColumns.emplace_back(i);
+        }
+        std::swap(taskMeta.ReadInfo.GroupByColumnNames, groupByColumns);
+    }
+    taskMeta.ReadInfo.ReadType = OlapReadTypeFromProto(readOlapRange->GetReadType());
+    taskMeta.ReadInfo.OlapProgram.Program = readOlapRange->GetOlapProgram();
+    for (auto& name: readOlapRange->GetOlapProgramParameterNames()) {
+        taskMeta.ReadInfo.OlapProgram.ParameterNames.insert(name);
+    }
+}
+
+void MergeReadInfoToTaskMeta(TTaskMeta& meta, ui64 shardId, TMaybe<TShardKeyRanges>& keyReadRanges,
+    const TPhysicalShardReadSettings& readSettings, const TVector<TTaskMeta::TColumn>& columns,
+    const NKqpProto::TKqpPhyTableOperation& op, bool isPersistentScan)
+{
+    TTaskMeta::TShardReadInfo readInfo = {
+        .Ranges = {},
+        .Columns = columns,
+    };
+    if (keyReadRanges) {
+        readInfo.Ranges = std::move(*keyReadRanges); // sorted & non-intersecting
+    }
+
+    if (isPersistentScan) {
+        readInfo.ShardId = shardId;
+    }
+
+    FillReadInfo(meta, readSettings.ItemsLimit, readSettings.GetSorting());
+    if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadOlapRange) {
+        FillOlapReadInfo(meta, readSettings.ResultType, op.GetReadOlapRange());
+    }
+
+    if (!meta.Reads) {
+        meta.Reads.ConstructInPlace();
+    }
+
+    meta.Reads->emplace_back(std::move(readInfo));
+}
+
+THashSet<ui64> TKqpTasksGraph::BuildDatashardTasks(TStageInfo& stageInfo, const IKqpTransactionManagerPtr& txManager) {
+    THashSet<ui64> shardsWithEffects;
+
+    THashMap<ui64, ui64> shardTasks; // shardId -> taskId
+    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    auto getShardTask = [&](ui64 shardId) -> TTask& {
+        YQL_ENSURE(!txManager);
+        auto it  = shardTasks.find(shardId);
+        if (it != shardTasks.end()) {
+            return GetTask(it->second);
+        }
+        auto& task = AddTask(stageInfo);
+        task.Meta.Type = TTaskMeta::TTaskType::DataShard;
+        task.Meta.ShardId = shardId;
+        shardTasks.emplace(shardId, task.Id);
+
+        return task;
+    };
+
+    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
+    const auto& keyTypes = tableInfo->KeyColumnTypes;
+
+    for (auto& op : stage.GetTableOps()) {
+        Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.TablePath == op.GetTable().GetPath());
+        switch (op.GetTypeCase()) {
+            case NKqpProto::TKqpPhyTableOperation::kUpsertRows:
+            case NKqpProto::TKqpPhyTableOperation::kDeleteRows: {
+                YQL_ENSURE(stage.InputsSize() <= 1, "Effect stage with multiple inputs: " << stage.GetProgramAst());
+
+                auto result = PartitionPruner->PruneEffect(op, stageInfo);
+                for (auto& [shardId, shardInfo] : result) {
+                    YQL_ENSURE(!shardInfo.KeyReadRanges);
+                    YQL_ENSURE(shardInfo.KeyWriteRanges);
+
+                    auto& task = getShardTask(shardId);
+
+                    if (!task.Meta.Writes) {
+                        task.Meta.Writes.ConstructInPlace();
+                        task.Meta.Writes->Ranges = std::move(*shardInfo.KeyWriteRanges);
+                    } else {
+                        task.Meta.Writes->Ranges.MergeWritePoints(std::move(*shardInfo.KeyWriteRanges), keyTypes);
+                    }
+
+                    if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kDeleteRows) {
+                        task.Meta.Writes->AddEraseOp();
+                    } else {
+                        task.Meta.Writes->AddUpdateOp();
+                    }
+
+                    for (const auto& [name, info] : shardInfo.ColumnWrites) {
+                        auto& column = tableInfo->Columns.at(name);
+
+                        auto& taskColumnWrite = task.Meta.Writes->ColumnWrites[column.Id];
+                        taskColumnWrite.Column.Id = column.Id;
+                        taskColumnWrite.Column.Type = column.Type;
+                        taskColumnWrite.Column.Name = name;
+                        taskColumnWrite.MaxValueSizeBytes = std::max(taskColumnWrite.MaxValueSizeBytes,
+                            info.MaxValueSizeBytes);
+                    }
+                    shardsWithEffects.insert(shardId);
+                }
+
+                break;
+            }
+
+            case NKqpProto::TKqpPhyTableOperation::kReadOlapRange: {
+                YQL_ENSURE(false, "The previous check did not work! Data query read does not support column shard tables." << Endl);
+                    // TODO: << this->DebugString());
+            }
+
+            default: {
+                YQL_ENSURE(false, "Unexpected table operation: " << (ui32) op.GetTypeCase() << Endl);
+                    // TODO: << this->DebugString());
+            }
+        }
+    }
+
+    // TODO: LOG_D("Stage " << stageInfo.Id << " will be executed on " << shardTasks.size() << " shards.");
+
+    // TODO:
+    // for (auto& shardTask : shardTasks) {
+    //     auto& task = GetTask(shardTask.second);
+    //     LOG_D("ActorState: " << CurrentStateFuncName()
+    //         << ", stage: " << stageInfo.Id << " create datashard task: " << shardTask.second
+    //         << ", shard: " << shardTask.first
+    //         << ", meta: " << task.Meta.ToString(keyTypes, *AppData()->TypeRegistry));
+    // }
+
+    return shardsWithEffects;
+}
+
+void PrepareScanMetaForUsage(TTaskMeta& meta, const TVector<NScheme::TTypeInfo>& keyTypes) {
+    YQL_ENSURE(meta.Reads.Defined());
+    auto& taskReads = meta.Reads.GetRef();
+
+    /*
+     * Sort read ranges so that sequential scan of that ranges produce sorted result.
+     *
+     * Partition pruner feed us with set of non-intersecting ranges with filled right boundary.
+     * So we may sort ranges based solely on the their rightmost point.
+     */
+    std::sort(taskReads.begin(), taskReads.end(), [&](const auto& lhs, const auto& rhs) {
+        if (lhs.ShardId == rhs.ShardId) {
+            return false;
+        }
+
+        const std::pair<const TSerializedCellVec*, bool> k1 = lhs.Ranges.GetRightBorder();
+        const std::pair<const TSerializedCellVec*, bool> k2 = rhs.Ranges.GetRightBorder();
+
+        const int cmp = CompareBorders<false, false>(
+            k1.first->GetCells(),
+            k2.first->GetCells(),
+            k1.second,
+            k2.second,
+            keyTypes);
+
+        return (cmp < 0);
+        });
+}
+
+ui32 TKqpTasksGraph::GetScanTasksPerNode(TStageInfo& stageInfo, const bool isOlapScan, const ui64 nodeId, bool enableShuffleElimination) const {
+    auto& intros = stageInfo.Introspections;
+    const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+    if (const auto taskCount = stage.GetTaskCount()) {
+        intros.push_back("Manually overridden - " + ToString(taskCount));
+        return taskCount;
+    }
+
+    ui32 result = 0;
+    if (isOlapScan) {
+        intros.push_back("This is OLAP scan");
+        if (AggregationSettings.HasCSScanThreadsPerNode()) {
+            result = AggregationSettings.GetCSScanThreadsPerNode();
+            intros.push_back("Using the CSScanThreadsPerNode value - " + ToString(result));
+        } else {
+            const TStagePredictor& predictor = stageInfo.Meta.Tx.Body->GetCalculationPredictor(stageInfo.Id.StageId);
+            result = predictor.CalcTasksOptimalCount(TStagePredictor::GetUsableThreads(), {}, intros);
+            intros.push_back("Predicted value for OLAP scan - " + ToString(result));
+        }
+    } else {
+        result = AggregationSettings.GetDSScanMinimalThreads();
+        intros.push_back("Using the DSScanMinimalThreads value - " + ToString(result));
+        if (stage.GetProgram().GetSettings().GetHasSort()) {
+            result = std::max(result, AggregationSettings.GetDSBaseSortScanThreads());
+            intros.push_back("Considering DSBaseSortScanThreads value because program has sort - " + ToString(result));
+        }
+        if (stage.GetProgram().GetSettings().GetHasMapJoin()) {
+            result = std::max(result, AggregationSettings.GetDSBaseJoinScanThreads());
+            intros.push_back("Considering DSBaseJoinScanThreads value because program has MapJoin - " + ToString(result));
+        }
+    }
+    result = Max<ui32>(1, result);
+
+    if (enableShuffleElimination) {
+        result *= 2;
+        intros.push_back("Multiply by 2 because of Shuffle Elimination - " + ToString(result));
+    }
+
+    intros.push_back("Predicted number of scan tasks per node " + ToString(nodeId) + " - " + ToString(result));
+    return result;
+}
+
+void TKqpTasksGraph::BuildScanTasksFromShards(TStageInfo& stageInfo, bool enableShuffleElimination, const TMap<ui64, ui64>& shardIdToNodeId,
+    TQueryExecutionStats* stats)
+{
+    THashMap<ui64, std::vector<ui64>> nodeTasks;
+    THashMap<ui64, std::vector<TShardInfoWithId>> nodeShards;
+    THashMap<ui64, ui64> assignedShardsCount;
+    auto& intros = stageInfo.Introspections;
+    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    auto& columnShardHashV1Params = stageInfo.Meta.ColumnShardHashV1Params;
+    bool shuffleEliminated = enableShuffleElimination && stage.GetIsShuffleEliminated();
+    if (shuffleEliminated && stageInfo.Meta.ColumnTableInfoPtr) {
+        const auto& tableDesc = stageInfo.Meta.ColumnTableInfoPtr->Description;
+        columnShardHashV1Params.SourceShardCount = tableDesc.GetColumnShardCount();
+        columnShardHashV1Params.SourceTableKeyColumnTypes = std::make_shared<TVector<NScheme::TTypeInfo>>();
+        for (const auto& column: tableDesc.GetSharding().GetHashSharding().GetColumns()) {
+            Y_ENSURE(stageInfo.Meta.TableConstInfo->Columns.contains(column), TStringBuilder{} << "Table doesn't have column: " << column);
+            auto columnType = stageInfo.Meta.TableConstInfo->Columns.at(column).Type;
+            columnShardHashV1Params.SourceTableKeyColumnTypes->push_back(columnType);
+        }
+    }
+
+    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
+    const auto& keyTypes = tableInfo->KeyColumnTypes;
+    for (auto& op : stage.GetTableOps()) {
+        Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.TablePath == op.GetTable().GetPath());
+
+        auto columns = BuildKqpColumns(op, tableInfo);
+        bool isFullScan;
+        auto partitions = PartitionPruner->Prune(op, stageInfo, isFullScan);
+        const bool isOlapScan = (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadOlapRange);
+        auto readSettings = ExtractReadSettings(op, stageInfo, TxAlloc->HolderFactory, TxAlloc->TypeEnv);
+
+        if (isFullScan && readSettings.ItemsLimit) {
+            Counters->Counters->FullScansExecuted->Inc();
+        }
+
+        if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadRange) {
+            stageInfo.Meta.SkipNullKeys.assign(op.GetReadRange().GetSkipNullKeys().begin(),
+                                                op.GetReadRange().GetSkipNullKeys().end());
+            // not supported for scan queries
+            YQL_ENSURE(!readSettings.IsReverse());
+        }
+
+        for (auto&& i: partitions) {
+            const ui64 nodeId = shardIdToNodeId.at(i.first);
+            nodeShards[nodeId].emplace_back(TShardInfoWithId(i.first, std::move(i.second)));
+        }
+
+        if (stats) {
+            for (auto&& i : nodeShards) {
+                stats->AddNodeShardsCount(stageInfo.Id.StageId, i.first, i.second.size());
+            }
+        }
+
+        if (!AppData()->FeatureFlags.GetEnableSeparationComputeActorsFromRead() && !shuffleEliminated || (!isOlapScan && readSettings.IsSorted())) {
+            THashMap<ui64 /* nodeId */, ui64 /* tasks count */> olapAndSortedTasksCount;
+
+            auto AssignScanTaskToShard = [&](const ui64 shardId, const bool sorted) -> TTask& {
+                ui64 nodeId = shardIdToNodeId.at(shardId);
+                if (stageInfo.Meta.IsOlap() && sorted) {
+                    auto& task = AddTask(stageInfo);
+                    task.Meta.NodeId = nodeId;
+                    task.Meta.ScanTask = true;
+                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
+                    ++olapAndSortedTasksCount[nodeId];
+                    return task;
+                }
+
+                auto& tasks = nodeTasks[nodeId];
+                auto& cnt = assignedShardsCount[nodeId];
+                const ui32 maxScansPerNode = GetScanTasksPerNode(stageInfo, isOlapScan, nodeId);
+                if (cnt < maxScansPerNode) {
+                    auto& task = AddTask(stageInfo);
+                    task.Meta.NodeId = nodeId;
+                    task.Meta.ScanTask = true;
+                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
+                    tasks.push_back(task.Id);
+                    ++cnt;
+                    return task;
+                } else {
+                    ui64 taskIdx = cnt % maxScansPerNode;
+                    ++cnt;
+                    intros.push_back("Scan task for node " + ToString(nodeId) + " not created");
+                    return GetTask(tasks[taskIdx]);
+                }
+            };
+
+            for (auto&& pair : nodeShards) {
+                auto& shardsInfo = pair.second;
+                for (auto&& shardInfo : shardsInfo) {
+                    auto& task = AssignScanTaskToShard(shardInfo.ShardId, readSettings.IsSorted());
+                    MergeReadInfoToTaskMeta(task.Meta, shardInfo.ShardId, shardInfo.KeyReadRanges, readSettings,
+                        columns, op, /*isPersistentScan*/ true);
+                }
+            }
+
+            for (const auto& pair : nodeTasks) {
+                for (const auto& taskIdx : pair.second) {
+                    auto& task = GetTask(taskIdx);
+                    task.Meta.SetEnableShardsSequentialScan(readSettings.IsSorted());
+                    PrepareScanMetaForUsage(task.Meta, keyTypes);
+                }
+
+                intros.push_back("Actual number of scan tasks for node " + ToString(pair.first) + " - " + ToString(pair.second.size()));
+            }
+
+            for (const auto& [nodeId, count] : olapAndSortedTasksCount) {
+                intros.push_back("Actual number of scan tasks (olap+sorted) for node " + ToString(nodeId) + " - " + ToString(count));
+            }
+        } else if (shuffleEliminated /* save partitioning for shuffle elimination */) {
+            std::size_t stageInternalTaskId = 0;
+            columnShardHashV1Params.TaskIndexByHash = std::make_shared<TVector<ui64>>();
+            columnShardHashV1Params.TaskIndexByHash->resize(columnShardHashV1Params.SourceShardCount);
+
+            for (auto&& pair : nodeShards) {
+                const auto nodeId = pair.first;
+                auto& shardsInfo = pair.second;
+                std::size_t maxTasksPerNode = std::min<std::size_t>(shardsInfo.size(), GetScanTasksPerNode(stageInfo, isOlapScan, nodeId, true));
+                std::vector<TTaskMeta> metas(maxTasksPerNode, TTaskMeta());
+                {
+                    for (std::size_t i = 0; i < shardsInfo.size(); ++i) {
+                        auto&& shardInfo = shardsInfo[i];
+                        MergeReadInfoToTaskMeta(
+                            metas[i % maxTasksPerNode],
+                            shardInfo.ShardId,
+                            shardInfo.KeyReadRanges,
+                            readSettings,
+                            columns, op,
+                            /*isPersistentScan*/ true
+                        );
+                    }
+
+                    for (auto& meta: metas) {
+                        PrepareScanMetaForUsage(meta, keyTypes);
+                        // TODO:
+                        // LOG_D("Stage " << stageInfo.Id << " create scan task meta for node: " << nodeId
+                        //     << ", meta: " << meta.ToString(keyTypes, *AppData()->TypeRegistry));
+                    }
+                }
+
+                // in runtime we calc hash, which will be in [0; shardcount]
+                // so we merge to mappings : hash -> shardID and shardID -> channelID for runtime
+                THashMap<ui64, ui64> hashByShardId;
+                Y_ENSURE(stageInfo.Meta.ColumnTableInfoPtr != nullptr, "ColumnTableInfoPtr is nullptr, maybe information about shards haven't beed delivered yet.");
+                const auto& tableDesc = stageInfo.Meta.ColumnTableInfoPtr->Description;
+                const auto& sharding = tableDesc.GetSharding();
+                for (std::size_t i = 0; i < sharding.ColumnShardsSize(); ++i) {
+                    hashByShardId.insert({sharding.GetColumnShards(i), i});
+                }
+
+                intros.push_back("Actual number of scan tasks from shards with shuffle elimination for node " + ToString(nodeId) + " - " + ToString(maxTasksPerNode));
+
+                for (ui32 t = 0; t < maxTasksPerNode; ++t, ++stageInternalTaskId) {
+                    auto& task = AddTask(stageInfo);
+                    task.Meta = metas[t];
+                    task.Meta.SetEnableShardsSequentialScan(false);
+                    task.Meta.NodeId = nodeId;
+                    task.Meta.ScanTask = true;
+                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
+                    task.SetMetaId(t);
+
+                    for (const auto& readInfo: *task.Meta.Reads) {
+                        Y_ENSURE(hashByShardId.contains(readInfo.ShardId));
+                        (*columnShardHashV1Params.TaskIndexByHash)[hashByShardId[readInfo.ShardId]] = stageInternalTaskId;
+                    }
+                }
+            }
+
+            LOG_DEBUG_S(
+                *TlsActivationContext,
+                NKikimrServices::KQP_EXECUTER,
+                "Stage with scan " << "[" << stageInfo.Id.TxId << ":" << stageInfo.Id.StageId << "]"
+                << " has keys: " << columnShardHashV1Params.KeyTypesToString() << " and task count: " << stageInternalTaskId;
+            );
+        } else {
+            ui32 metaId = 0;
+            for (auto&& pair : nodeShards) {
+                const auto nodeId = pair.first;
+                auto& shardsInfo = pair.second;
+                const ui32 metaGlueingId = ++metaId;
+                TTaskMeta meta;
+                {
+                    for (auto&& shardInfo : shardsInfo) {
+                        MergeReadInfoToTaskMeta(meta, shardInfo.ShardId, shardInfo.KeyReadRanges, readSettings,
+                            columns, op, /*isPersistentScan*/ true);
+                    }
+                    PrepareScanMetaForUsage(meta, keyTypes);
+                    // TODO:
+                    // LOG_D("Stage " << stageInfo.Id << " create scan task meta for node: " << nodeId
+                    //     << ", meta: " << meta.ToString(keyTypes, *AppData()->TypeRegistry));
+                }
+
+                const auto maxTasksPerNode = GetScanTasksPerNode(stageInfo, isOlapScan, nodeId);
+                intros.push_back("Actual number of scan tasks from shards without shuffle elimination for node " + ToString(nodeId) + " - " + ToString(maxTasksPerNode));
+
+                for (ui32 t = 0; t < maxTasksPerNode; ++t) {
+                    auto& task = AddTask(stageInfo);
+                    task.Meta = meta;
+                    task.Meta.SetEnableShardsSequentialScan(false);
+                    task.Meta.NodeId = nodeId;
+                    task.Meta.ScanTask = true;
+                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
+                    task.SetMetaId(metaGlueingId);
+                }
+            }
+        }
+    }
+
+    // TODO: LOG_D("Stage " << stageInfo.Id << " will be executed on " << nodeTasks.size() << " nodes.");
+}
+
+void FillReadTaskFromSource(TTask& task, const TString& sourceName, const TString& structuredToken, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui64 nodeOffset) {
+    if (structuredToken) {
+        task.Meta.SecureParams.emplace(sourceName, structuredToken);
+    }
+
+    if (resourceSnapshot.empty()) {
+        task.Meta.Type = TTaskMeta::TTaskType::Compute;
+    } else {
+        task.Meta.NodeId = resourceSnapshot[nodeOffset % resourceSnapshot.size()].GetNodeId();
+        task.Meta.Type = TTaskMeta::TTaskType::Scan;
+    }
+}
+
+void TKqpTasksGraph::BuildReadTasksFromSource(TStageInfo& stageInfo, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot,
+    ui32 scheduledTaskCount, const std::map<TString, TString>& secureParams)
+{
+    auto& intros = stageInfo.Introspections;
+    const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    YQL_ENSURE(stage.GetSources(0).HasExternalSource());
+    YQL_ENSURE(stage.SourcesSize() == 1, "multiple sources in one task are not supported");
+
+    const auto& stageSource = stage.GetSources(0);
+    const auto& externalSource = stageSource.GetExternalSource();
+
+    ui32 taskCount = externalSource.GetPartitionedTaskParams().size();
+    intros.push_back("Using number of PartitionedTaskParams from external source - " + ToString(taskCount));
+
+    auto taskCountHint = stage.GetTaskCount();
+    TString introHint;
+    if (taskCountHint == 0) {
+        taskCountHint = scheduledTaskCount;
+        introHint = "(Using scheduled task count as hint for override - " + ToString(taskCountHint) + ")";
+    }
+
+    if (taskCountHint) {
+        if (taskCount > taskCountHint) {
+            taskCount = taskCountHint;
+            if (!introHint.empty()) {
+                intros.push_back(introHint);
+            }
+            intros.push_back("Manually overridden - " + ToString(taskCount));
+        }
+    } else if (!resourceSnapshot.empty()) {
+        ui32 maxTaskcount = resourceSnapshot.size() * 2;
+        if (taskCount > maxTaskcount) {
+            taskCount = maxTaskcount;
+            intros.push_back("Using less tasks because of resource snapshot size - " + ToString(taskCount));
+        }
+    }
+
+    auto sourceName = externalSource.GetSourceName();
+    TString structuredToken;
+    if (sourceName) {
+        structuredToken = NYql::CreateStructuredTokenParser(externalSource.GetAuthInfo()).ToBuilder().ReplaceReferences(secureParams).ToJson();
+    }
+
+    ui64 nodeOffset = 0;
+    for (size_t i = 0; i < resourceSnapshot.size(); ++i) {
+        if (resourceSnapshot[i].GetNodeId() == GetMeta().ExecuterId.NodeId()) {
+            nodeOffset = i;
+            break;
+        }
+    }
+
+    if (GetMeta().IsRestored) {
+        for (const auto taskId : stageInfo.Tasks) {
+            auto& task = GetTask(taskId);
+            FillReadTaskFromSource(task, sourceName, structuredToken, resourceSnapshot, nodeOffset++);
+        }
+        return;
+    }
+
+    TVector<ui64> tasksIds;
+    tasksIds.reserve(taskCount);
+
+    // generate all tasks
+    for (ui32 i = 0; i < taskCount; i++) {
+        auto& task = AddTask(stageInfo);
+
+        if (!externalSource.GetEmbedded()) {
+            auto& input = task.Inputs[stageSource.GetInputIndex()];
+            input.ConnectionInfo = NYql::NDq::TSourceInput{};
+            input.SourceSettings = externalSource.GetSettings();
+            input.SourceType = externalSource.GetType();
+        }
+
+        FillReadTaskFromSource(task, sourceName, structuredToken, resourceSnapshot, nodeOffset++);
+
+        tasksIds.push_back(task.Id);
+    }
+
+    // distribute read ranges between them
+    ui32 currentTaskIndex = 0;
+    for (const TString& partitionParam : externalSource.GetPartitionedTaskParams()) {
+        GetTask(tasksIds[currentTaskIndex]).Meta.ReadRanges.push_back(partitionParam);
+        if (++currentTaskIndex >= tasksIds.size()) {
+            currentTaskIndex = 0;
+        }
+    }
+}
+
+TVector<TVector<TShardRangesWithShardId>> DistributeShardsToTasks(TVector<TShardRangesWithShardId> shardsRanges, const size_t tasksCount, const TVector<NScheme::TTypeInfo>& keyTypes) {
+    // TODO:
+    // if (IsDebugLogEnabled()) {
+    //     TStringBuilder sb;
+    //     sb << "Distributing shards to tasks: [";
+    //     for(size_t i = 0; i < shardsRanges.size(); i++) {
+    //         sb << "# " << i << ": " << shardsRanges[i].Ranges->ToString(keyTypes, *AppData()->TypeRegistry);
+    //     }
+
+    //     sb << " ].";
+    //     LOG_D(sb);
+    // }
+
+    std::sort(std::begin(shardsRanges), std::end(shardsRanges), [&](const TShardRangesWithShardId& lhs, const TShardRangesWithShardId& rhs) {
+            return CompareBorders<true, true>(
+                lhs.Ranges->GetRightBorder().first->GetCells(),
+                rhs.Ranges->GetRightBorder().first->GetCells(),
+                lhs.Ranges->GetRightBorder().second,
+                rhs.Ranges->GetRightBorder().second,
+                keyTypes) < 0;
+        });
+
+    // One shard (ranges set) can be assigned only to one task. Otherwise, we can break some optimizations like removing unnecessary shuffle.
+    TVector<TVector<TShardRangesWithShardId>> result(tasksCount);
+    size_t shardIndex = 0;
+    for (size_t taskIndex = 0; taskIndex < tasksCount; ++taskIndex) {
+        const size_t tasksLeft = tasksCount - taskIndex;
+        const size_t shardsLeft = shardsRanges.size() - shardIndex;
+        const size_t shardsPerCurrentTask = (shardsLeft + tasksLeft - 1) / tasksLeft;
+
+        for (size_t currentShardIndex = 0; currentShardIndex < shardsPerCurrentTask; ++currentShardIndex, ++shardIndex) {
+            result[taskIndex].push_back(shardsRanges[shardIndex]);
+        }
+    }
+    return result;
+}
+
+TMaybe<size_t> TKqpTasksGraph::BuildScanTasksFromSource(TStageInfo& stageInfo, bool limitTasksPerNode,
+    const TMap<ui64, ui64>& shardIdToNodeId, TQueryExecutionStats* stats)
+{
+    auto& intros = stageInfo.Introspections;
+    THashMap<ui64, std::vector<ui64>> nodeTasks;
+    THashMap<ui64, ui64> assignedShardsCount;
+
+    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+
+    bool singlePartitionedStage = stage.GetIsSinglePartition();
+
+    YQL_ENSURE(stage.GetSources(0).HasReadRangesSource());
+    YQL_ENSURE(stage.GetSources(0).GetInputIndex() == 0 && stage.SourcesSize() == 1);
+    for (auto& input : stage.inputs()) {
+        YQL_ENSURE(input.HasBroadcast());
+    }
+
+    auto& source = stage.GetSources(0).GetReadRangesSource();
+
+    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
+    const auto& keyTypes = tableInfo->KeyColumnTypes;
+
+    YQL_ENSURE(tableInfo->TableKind != NKikimr::NKqp::ETableKind::Olap);
+
+    auto columns = BuildKqpColumns(source, tableInfo);
+
+    const auto& snapshot = GetMeta().Snapshot;
+
+    TVector<ui64> createdTasksIds;
+    auto createNewTask = [&](
+            TMaybe<ui64> nodeId,
+            ui64 taskLocation,
+            TMaybe<ui64> shardId,
+            TMaybe<ui64> maxInFlightShards) -> TTask& {
+        auto& task = AddTask(stageInfo);
+        task.Meta.Type = TTaskMeta::TTaskType::Scan;
+        if (nodeId) {
+            task.Meta.NodeId = *nodeId;
+        }
+
+        if (!nodeId || !GetMeta().ShardsResolved) {
+            YQL_ENSURE(!GetMeta().ShardsResolved);
+            task.Meta.ShardId = taskLocation;
+        }
+
+        const auto& stageSource = stage.GetSources(0);
+        auto& input = task.Inputs[stageSource.GetInputIndex()];
+        input.SourceType = NYql::KqpReadRangesSourceName;
+        input.ConnectionInfo = NYql::NDq::TSourceInput{};
+
+        // allocating source settings
+
+        input.Meta.SourceSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpReadRangesSourceSettings>();
+        NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
+        FillTableMeta(stageInfo, settings->MutableTable());
+
+        settings->SetIsTableImmutable(source.GetIsTableImmutable());
+
+        for (auto& keyColumn : keyTypes) {
+            auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(keyColumn, "");
+            *settings->AddKeyColumnTypeInfos() = columnType.TypeInfo ?
+                *columnType.TypeInfo :
+                NKikimrProto::TTypeInfo();
+            settings->AddKeyColumnTypes(static_cast<ui32>(keyColumn.GetTypeId()));
+        }
+
+        for (auto& column : columns) {
+            auto* protoColumn = settings->AddColumns();
+            protoColumn->SetId(column.Id);
+            auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(column.Type, column.TypeMod);
+            protoColumn->SetType(columnType.TypeId);
+            protoColumn->SetNotNull(column.NotNull);
+            protoColumn->SetIsPrimary(column.IsPrimary);
+            if (columnType.TypeInfo) {
+                *protoColumn->MutableTypeInfo() = *columnType.TypeInfo;
+            }
+            protoColumn->SetName(column.Name);
+        }
+
+        if (GetMeta().CheckDuplicateRows) {
+            for (auto& colName : tableInfo->KeyColumns) {
+                const auto& tableColumn = tableInfo->Columns.at(colName);
+                auto* protoColumn = settings->AddDuplicateCheckColumns();
+                protoColumn->SetId(tableColumn.Id);
+                auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(tableColumn.Type, tableColumn.TypeMod);
+                protoColumn->SetType(columnType.TypeId);
+                if (columnType.TypeInfo) {
+                    *protoColumn->MutableTypeInfo() = *columnType.TypeInfo;
+                }
+                protoColumn->SetName(colName);
+            }
+        }
+
+        if (AppData()->FeatureFlags.GetEnableArrowFormatAtDatashard()) {
+            settings->SetDataFormat(NKikimrDataEvents::FORMAT_ARROW);
+        } else {
+            settings->SetDataFormat(NKikimrDataEvents::FORMAT_CELLVEC);
+        }
+
+        if (snapshot.IsValid()) {
+            settings->MutableSnapshot()->SetStep(snapshot.Step);
+            settings->MutableSnapshot()->SetTxId(snapshot.TxId);
+        }
+
+        if (GetMeta().RequestIsolationLevel == NKikimrKqp::ISOLATION_LEVEL_READ_UNCOMMITTED) {
+            settings->SetAllowInconsistentReads(true);
+        }
+
+        settings->SetReverse(source.GetReverse());
+        settings->SetSorted(source.GetSorted());
+
+        if (maxInFlightShards) {
+            settings->SetMaxInFlightShards(*maxInFlightShards);
+        }
+
+        if (shardId) {
+            settings->SetShardIdHint(*shardId);
+        }
+
+        if (GetMeta().MaxBatchSize) {
+            settings->SetItemsLimit(*GetMeta().MaxBatchSize);
+            settings->SetIsBatch(true);
+        } else {
+            ui64 itemsLimit = ExtractItemsLimit(stageInfo, source.GetItemsLimit(), TxAlloc->HolderFactory, TxAlloc->TypeEnv);
+            settings->SetItemsLimit(itemsLimit);
+        }
+
+        auto& lockTxId = GetMeta().LockTxId;
+        if (lockTxId) {
+            settings->SetLockTxId(*lockTxId);
+            settings->SetLockNodeId(GetMeta().ExecuterId.NodeId());
+        }
+
+        if (GetMeta().LockMode) {
+            settings->SetLockMode(*GetMeta().LockMode);
+        }
+
+        createdTasksIds.push_back(task.Id);
+        return task;
+    };
+
+    THashMap<ui64, TVector<ui64>> nodeIdToTasks;
+    THashMap<ui64, TVector<TShardRangesWithShardId>> nodeIdToShardKeyRanges;
+    Y_DEFER {
+        intros.push_back("Built scan tasks from source and shards to read for node");
+        for (const auto& [nodeId, tasks] : nodeIdToTasks) {
+            intros.push_back(ToString(nodeId) + " - " + ToString(tasks.size()) + ", " + ToString(nodeIdToShardKeyRanges.at(nodeId).size()));
+        }
+        intros.push_back("Total built scan tasks from source - " + ToString(createdTasksIds.size()));
+    };
+
+    auto addPartition = [&](
+        ui64 taskLocation,
+        TMaybe<ui64> nodeId,
+        TMaybe<ui64> shardId,
+        const TShardInfo& shardInfo,
+        TMaybe<ui64> maxInFlightShards = Nothing())
+    {
+        YQL_ENSURE(!shardInfo.KeyWriteRanges);
+
+        if (!nodeId) {
+            const auto nodeIdPtr = shardIdToNodeId.FindPtr(taskLocation);
+            nodeId = nodeIdPtr ? TMaybe<ui64>{*nodeIdPtr} : Nothing();
+        }
+
+        YQL_ENSURE(!GetMeta().ShardsResolved || nodeId);
+
+        YQL_ENSURE(stats);
+        if (shardId) {
+            stats->AffectedShards.insert(*shardId);
+        }
+
+        if (limitTasksPerNode && GetMeta().ShardsResolved) {
+            const auto maxScanTasksPerNode = GetScanTasksPerNode(stageInfo, /* isOlapScan */ false, *nodeId);
+            auto& nodeTasks = nodeIdToTasks[*nodeId];
+            if (nodeTasks.size() < maxScanTasksPerNode) {
+                const auto& task = createNewTask(nodeId, taskLocation, {}, maxInFlightShards);
+                nodeTasks.push_back(task.Id);
+            }
+
+            nodeIdToShardKeyRanges[*nodeId].push_back(TShardRangesWithShardId{shardId, &*shardInfo.KeyReadRanges});
+        } else {
+            auto& task = createNewTask(nodeId, taskLocation, shardId, maxInFlightShards);
+            const auto& stageSource = stage.GetSources(0);
+            auto& input = task.Inputs[stageSource.GetInputIndex()];
+            NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
+
+            shardInfo.KeyReadRanges->SerializeTo(settings);
+        }
+    };
+
+    auto fillRangesForTasks = [&]() {
+        for (const auto& [nodeId, shardsRanges] : nodeIdToShardKeyRanges) {
+            const auto& tasks = nodeIdToTasks.at(nodeId);
+
+            const auto rangesDistribution = DistributeShardsToTasks(shardsRanges, tasks.size(), keyTypes);
+            YQL_ENSURE(rangesDistribution.size() == tasks.size());
+
+            for (size_t taskIndex = 0; taskIndex < tasks.size(); ++taskIndex) {
+                const auto taskId = tasks[taskIndex];
+                auto& task = GetTask(taskId);
+                const auto& stageSource = stage.GetSources(0);
+                auto& input = task.Inputs[stageSource.GetInputIndex()];
+                NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
+
+                const auto& shardsRangesForTask = rangesDistribution[taskIndex];
+
+                if (shardsRangesForTask.size() == 1 && shardsRangesForTask[0].ShardId) {
+                    settings->SetShardIdHint(*shardsRangesForTask[0].ShardId);
+                }
+
+                bool hasRanges = false;
+                for (const auto& shardRanges : shardsRangesForTask) {
+                    hasRanges |= shardRanges.Ranges->HasRanges();
+                }
+
+                for (const auto& shardRanges : shardsRangesForTask) {
+                    shardRanges.Ranges->SerializeTo(settings, !hasRanges);
+                }
+            }
+        }
+    };
+
+    bool isFullScan = false;
+    const THashMap<ui64, TShardInfo> partitions = GetMeta().SourceScanStageIdToParititions.empty()
+        ? PartitionPruner->Prune(source, stageInfo, isFullScan)
+        : GetMeta().SourceScanStageIdToParititions.at(stageInfo.Id);
+
+    if (isFullScan && !source.HasItemsLimit()) {
+        Counters->Counters->FullScansExecuted->Inc();
+    }
+
+    bool isSequentialInFlight = source.GetSequentialInFlightShards() > 0 && partitions.size() > source.GetSequentialInFlightShards();
+    bool isParallelPointRead = GetMeta().EnableParallelPointReadConsolidation && !isSequentialInFlight && !source.GetSorted() && IsParallelPointReadPossible(partitions);
+
+    if (partitions.size() > 0 && (isSequentialInFlight || isParallelPointRead || singlePartitionedStage)) {
+        auto [startShard, shardInfo] = PartitionPruner->MakeVirtualTablePartition(source, stageInfo);
+
+        YQL_ENSURE(stats);
+        for (auto& [shardId, _] : partitions) {
+            stats->AffectedShards.insert(shardId);
+        }
+
+        TMaybe<ui64> inFlightShards = Nothing();
+        if (isSequentialInFlight) {
+            inFlightShards = source.GetSequentialInFlightShards();
+        }
+
+        if (shardInfo.KeyReadRanges) {
+            const TMaybe<ui64> nodeId = (isParallelPointRead || singlePartitionedStage) ? TMaybe<ui64>{GetMeta().ExecuterId.NodeId()} : Nothing();
+            addPartition(startShard, nodeId, {}, shardInfo, inFlightShards);
+            fillRangesForTasks();
+            return (isParallelPointRead || singlePartitionedStage) ? TMaybe<size_t>(partitions.size()) : Nothing();
+        } else {
+            return 0;
+        }
+    } else {
+        for (auto& [shardId, shardInfo] : partitions) {
+            addPartition(shardId, {}, shardId, shardInfo, {});
+        }
+        fillRangesForTasks();
+        return partitions.size();
+    }
+}
+
+TKqpTasksGraph::TKqpTasksGraph(const NKikimr::NKqp::TTxAllocatorState::TPtr& txAlloc,
+    const TPartitionPrunerConfig& partitionPrunerConfig,
+    const NKikimrConfig::TTableServiceConfig::TAggregationConfig& aggregationSettings,
+    const TKqpRequestCounters::TPtr& counters)
+    : PartitionPruner(MakeHolder<TPartitionPruner>(txAlloc->HolderFactory, txAlloc->TypeEnv, std::move(partitionPrunerConfig)))
+    , TxAlloc(txAlloc)
+    , AggregationSettings(aggregationSettings)
+    , Counters(counters)
+{}
 
 TString TTaskMeta::ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const NScheme::TTypeRegistry& typeRegistry) const
 {
@@ -1700,5 +2741,4 @@ TString TTaskMeta::ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const N
     return sb;
 }
 
-} // namespace NKqp
-} // namespace NKikimr
+} // namespace NKikimr::NKqp
